requirements.txt
# Core RAG dependencies
langchain==0.3.15
langchain-community==0.3.15
langchain-aws==0.2.14
langchain-pinecone==0.2.0

# Vector store
pinecone-client==5.0.1

# AWS Bedrock (compatible con langchain-aws 0.2.14)
boto3>=1.37.0
botocore>=1.40.0

# Document processing
pypdf==4.3.1
python-dotenv==1.0.1

# LangSmith monitoring
langsmith==0.1.143

# Utilities
pydantic==2.10.5
pydantic-settings==2.7.1